+++ Remote2D/src/com/remote/remote2d/editor/inspector/GuiEditorInspectorSectionInt.java
@@ -49,7 +49,7 @@ public class GuiEditorInspectorSectionInt extends GuiEditorInspectorSection {
 	public void setData(Object o) {
 		if(o instanceof Integer)
 		{
-			textField.text = (Integer)o+"";
+			textField.text = o+"";
 		}
 	}
